set targetSpeed type from uint to int

+++ bbb-car-controller/lib/controller_statemachine/controller_statemachine.hpp
@@ -76,7 +76,7 @@ typedef struct{
     evCarEvent speed;
     evCameraEvent camera;
     evGoStatusEvent goStatus;
-    uint8_t targetSpeed;
+    int targetSpeed;
     uint8_t direction; // 1 - left, 0 - straight, 2 - right
 } stCarStatemachine;
